update project `dbft` structure
