+++ src/DBFTPlugin/Types/ChangeViewReason.cs
