+++ src/DBFTPlugin/Types/ConsensusMessageType.cs
